+++ changelog.md
@@ -1,5 +1,10 @@
 # changelog
 
+## 0.52.3
+
+- **break**: register schemas for gen and search for them before the file resolver
+  ([#316](https://github.com/feltcoop/gro/pull/316))
+
 ## 0.52.2
 
 - improve safety of `gro deploy` and update its docs